+++ packages/config-tailwind/src/plugin/base.plugin.ts
@@ -11,17 +11,26 @@ const base = plugin(({ addBase }) => {
     '::selection': {
       '@apply bg-primary text-background': '',
     },
-    'input:-webkit-autofill, textarea:-webkit-autofill, select:-webkit-autofill':
-      {
-        '-webkit-box-shadow':
-          '0 0 0 1000px hsl(var(--background)) inset !important',
-        '-webkit-text-fill-color': 'hsl(var(--foreground)) !important',
-        'box-shadow': '0 0 0 1000px hsl(var(--background)) inset !important',
+    'input, textarea, select': {
+      '&:-webkit-autofill': {
+        '&, &:hover, &:focus': {
+          'box-shadow': 'inset 0 0 0 1000px hsl(var(--background))',
+          '-webkit-box-shadow': 'inset 0 0 0 1000px hsl(var(--background)) ',
+          '-webkit-text-fill-color': 'hsl(var(--foreground))',
+        },
       },
-    "button, [role='button'], input:where([type='button']), input:where([type='reset']), input:where([type='submit'])":
-      {
+    },
+    "button, [role='button']": {
+      '@apply outline-transparent': '',
+    },
+    input: {
+      "&[type='button'], &[type='reset'], &[type='submit']": {
         '@apply outline-transparent': '',
       },
+      '&::-webkit-search-cancel-button, &::-webkit-search-decoration': {
+        '@apply appearance-none': '',
+      },
+    },
   });
 });