+++ packages/primitive/number-input/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @codefast-ui/number-input
 
+## 0.1.29
+
+### Patch Changes
+
+- Update Tailwind config for webkit autofill and buttons
+- Updated dependencies
+  - @codefast-ui/input@0.1.29
+
 ## 0.1.28
 
 ### Patch Changes